Refactor: clean up the utils.
- Keep the way responses are handled consistent.
- Fix some typos
- Move the general utils to `utils/index.js` instead of `utils/gitApi.js`
