+++ frontend/src/utils/giteaApi.js
@@ -1,5 +1,6 @@
-import path from 'path'
 import slugify from 'slugify'
+
+import { urlToName } from "@utils/index";
 
 const giteaApiUrl = `${process.env.KITSPACE_GITEA_URL}/api/v1`
 const credentials = 'include'
@@ -68,7 +69,7 @@ export const updateRepo = async (repo, updateFields, csrf) => {
  * @param csrf {string}
  * @returns {Promise<Response>}
  */
-export const migrateRepo = async (remoteRepo, uid, csrf) => {
+export const mirrorRepo = async (remoteRepo, uid, csrf) => {
   const clone_addr = remoteRepo
   const repo_name = urlToName(clone_addr)
   const endpoint = `${giteaApiUrl}/repos/migrate?_csrf=${csrf}`
@@ -92,7 +93,7 @@ export const migrateRepo = async (remoteRepo, uid, csrf) => {
   })
 }
 /**
- * delete the corresponding gitea repo for a project.
+ * Delete the corresponding gitea repo for a project.
  * @param repo {string}
  * @param csrf {string}
  * @returns {Promise<boolean>}
@@ -116,11 +117,7 @@ export const deleteRepo = async (repo, csrf) => {
  * @param branch {string=}
  * @returns {Promise<Array|null>}
  */
-export const getRepoFiles = async (repo, csrf, branch) => {
-  if (!branch) {
-    branch = 'master'
-  }
-
+export const getRepoFiles = async (repo, csrf, branch = 'master') => {
   const endpoint = `${giteaApiUrl}/repos/${repo}/contents?ref=${branch}&_csrf=${csrf}`
   const res = await fetch(endpoint, {
     method: 'GET',
@@ -129,12 +126,17 @@ export const getRepoFiles = async (repo, csrf, branch) => {
     headers,
   })
 
-  const body = await res.json()
+  if (res.ok) {
+    const body = await res.json()
 
-  if (body.hasOwnProperty('owner')) {
-    return []
+    // For some reason if the repo is empty the gitea api returns the repo details instead of empty array!
+    if (body.hasOwnProperty('owner')) {
+      return []
+    } else {
+      return body
+    }
   } else {
-    return res.ok ? body : []
+    return []
   }
 }
 
@@ -173,14 +175,14 @@ export const getRepo = async fullname => {
  * @param csrf{string}
  * @returns {Promise<any|*[]>}
  */
-export const getRepos = csrf => searchRepos(csrf)
+export const getAllRepos = csrf => searchRepos(csrf)
 
 /**
  * Search all repos
  * @param csrf{string}
  * @param sort{string}
  * @param order{string}
- * @param q{string}: search query, leave undefined to return all repos
+ * @param q{string=}: search query, leave undefined to return all repos
  * @returns {Promise<any|*[]>}
  */
 export const searchRepos = async (csrf, sort = 'updated', order = 'desc', q) => {
@@ -194,13 +196,18 @@ export const searchRepos = async (csrf, sort = 'updated', order = 'desc', q) => 
     sort,
     order,
     q,
-  }).then(r => r.json())
+  })
 
-  return res.ok ? res.data : []
+  if (res.ok) {
+    const body = await res.json()
+    return body.data
+  } else {
+    return []
+  }
 }
 
 /**
- * Get the repos own by the current authenticated user
+ * Get the repos owned by the current authenticated user.
  * @param csrf{string}
  * @returns {Promise<any|*[]>}
  */
@@ -256,20 +263,4 @@ export const uploadFile = async (repo, path, content, csrf) => {
   console.log(body)
 
   return res.ok
-}
-
-export const urlToName = url => {
-  url = new URL(url)
-  return path.basename(url.pathname, path.extname(url.pathname))
-}
-export const projectNameFromPath = path => {
-  return path.split('/').slice(3).join('/')
-}
-export const getSession = req => {
-  if (req != null) {
-    return req.session
-  }
-  if (process.browser) {
-    return window.session
-  }
 }
